# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediCare HMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bills:
    get:
      operationId: listBills
      tags: [billing]
      parameters:
        - { in: query, name: patientId, schema: { type: integer } }
        - { in: query, name: status, schema: { type: string } }
      responses:
        "200":
          description: Bills
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bill" }
    post:
      operationId: createBill
      tags: [billing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bill" }
  /bills/{id}:
    get:
      operationId: getBill
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bill" }
  /bills/{id}/pay:
    post:
      operationId: payBill
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Paid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bill" }
  /bills/{id}/full:
    get:
      operationId: getBillFull
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Bill with payments and refunds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillFull" }
  /bills/{id}/payments:
    get:
      operationId: listBillPayments
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
    post:
      operationId: recordPayment
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
  /bills/{id}/refunds:
    post:
      operationId: recordRefund
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefundInput" }
      responses:
        "201":
          description: Refund recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }
  /bills/{id}/void:
    post:
      operationId: voidBill
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidInput" }
      responses:
        "200":
          description: Voided
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bill" }
  /bills/{id}/claim:
    post:
      operationId: updateBillClaim
      tags: [billing]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClaimUpdateInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bill" }
...
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendNotificationInput" }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationLog" }
...
components:
  schemas:
...
    Bill:
      type: object
      required: [id, patientId, patientName, billNumber, subtotal, discount, cgst, sgst, igst, total, paidAmount, refundedAmount, balance, status, gstMode, claimStatus, claimAmount, createdAt, items]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        patientPhone: { type: ["string", "null"] }
        patientEmail: { type: ["string", "null"] }
        doctorId: { type: ["integer", "null"] }
        doctorName: { type: ["string", "null"] }
        department: { type: ["string", "null"] }
        billNumber: { type: string }
        subtotal: { type: number }
        discount: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        total: { type: number }
        paidAmount: { type: number }
        refundedAmount: { type: number }
        balance: { type: number }
        status: { type: string }
        gstMode: { type: string }
        paymentMethod: { type: ["string", "null"] }
        insuranceProvider: { type: ["string", "null"] }
        tpa: { type: ["string", "null"] }
        policyNumber: { type: ["string", "null"] }
        preAuthCode: { type: ["string", "null"] }
        claimStatus: { type: string }
        claimAmount: { type: number }
        notes: { type: ["string", "null"] }
        voidedAt: { type: ["string", "null"] }
        voidReason: { type: ["string", "null"] }
        createdAt: { type: string }
        paidAt: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/BillItem" }
    BillItem:
      type: object
      required: [description, quantity, unitPrice, amount]
      properties:
        serviceCode: { type: string }
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        discount: { type: number }
        gstRate: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        amount: { type: number }
    BillInput:
      x-doctor-note: doctorId/department are optional and used for billing reports.
      type: object
      required: [patientId, items]
      properties:
        patientId: { type: integer }
        doctorId: { type: integer }
        department: { type: string }
        gstMode: { type: string }
        discount: { type: number }
        insuranceProvider: { type: string }
        tpa: { type: string }
        policyNumber: { type: string }
        preAuthCode: { type: string }
        notes: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/BillItem" }
    Payment:
      type: object
      required: [id, billId, receiptNumber, amount, mode, receivedAt]
      properties:
        id: { type: integer }
        billId: { type: integer }
        receiptNumber: { type: string }
        amount: { type: number }
        tenderedAmount: { type: ["number", "null"] }
        changeDue: { type: ["number", "null"] }
        mode: { type: string }
        reference: { type: ["string", "null"] }
        receivedBy: { type: ["string", "null"] }
        cashierSessionId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        receivedAt: { type: string }
    PaymentInput:
      type: object
      required: [amount, mode]
      properties:
        amount: { type: number, description: "amount applied to the bill (≤ outstanding balance)" }
        tenderedAmount: { type: number, description: "what the customer handed over; for cash overpay, change = tendered − amount" }
        mode: { type: string }
        reference: { type: string }
        notes: { type: string }
    Refund:
      type: object
      required: [id, billId, amount, mode, reason, refundedAt]
      properties:
        id: { type: integer }
        billId: { type: integer }
        paymentId: { type: ["integer", "null"] }
        amount: { type: number }
        mode: { type: string }
        reason: { type: string }
        approvedBy: { type: ["string", "null"] }
        refundedAt: { type: string }
    RefundInput:
      type: object
      required: [amount, reason, mode]
      properties:
        paymentId: { type: integer }
        amount: { type: number }
        mode: { type: string }
        reason: { type: string }
    VoidInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    ClaimUpdateInput:
      type: object
      required: [claimStatus]
      properties:
        claimStatus: { type: string }
        claimAmount: { type: number }
        tpa: { type: string }
        policyNumber: { type: string }
        preAuthCode: { type: string }
    BillFull:
      type: object
      required: [bill, payments, refunds]
      properties:
        bill: { $ref: "#/components/schemas/Bill" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/Payment" }
        refunds:
          type: array
          items: { $ref: "#/components/schemas/Refund" }
...
    SendNotificationInput:
      type: object
      required: [eventKey, channel, patientId, variables]
      properties:
        eventKey: { type: string }
        channel: { type: string }
        patientId: { type: integer }
        recipientPhone: { type: string }
        variables:
          type: object
          additionalProperties: { type: string }
    NotificationLog:
      type: object
      required: [id, eventKey, channel, status, renderedBody, recipientPhone, sentAt]
      properties:
        id: { type: integer }
        patientId: { type: ["integer", "null"] }
        patientName: { type: ["string", "null"] }
        eventKey: { type: string }
        channel: { type: string }
        templateId: { type: ["integer", "null"] }
        renderedBody: { type: string }
        recipientPhone: { type: string }
        status: { type: string }
        providerRef: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        sentAt: { type: string }
...